+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 0.2.1
+
+- Add detailed help for commands
+- Add `!help` and `?` for usage in shell
+- Fix crashes on unexpected input
+- Fix exception handling
+
 ## Version 0.2.0
 
 - Renamed to TrainSH